% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdryfcn.R
\name{bdryfcn}
\alias{bdryfcn}
\title{Stopping Rule Boundary Function}
\usage{
bdryfcn(data.type, ...)
}
\arguments{
\item{data.type}{Type of data, choices include 'bin' for binary data and 'surv' for time-to-event data}

\item{...}{Other options to be passed to the corresponding stopping rule calculation. Please refer to the corresponding data type-specific \code{bdryfcn()} function for more details}
}
\value{
A univariate function that defines the rejection boundary at any number of evaluable patients (binary data) or amount of follow-up time (time-to-event data)
}
\description{
A wrapper function to calculate the boundary for a given stopping rule for safety monitoring for time-to-event data or binary data
}
