% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.bnd.s.R
\name{calc.bnd.s}
\alias{calc.bnd.s}
\title{Stopping Boundary Calculation (Survival Data)}
\usage{
calc.bnd.s(n, tau, p0, cval, type, param = NULL)
}
\arguments{
\item{n}{Maximum sample size for safety monitoring}

\item{tau}{Observation period}

\item{p0}{The toxicity rate under the null hypothesis}

\item{cval}{Critical value for stopping rule method}

\item{type}{The method used for constructing the stopping rule}

\item{param}{Extra parameter(s) needed for certain stopping rule methods. For Wang-Tsiatis test, this is the Delta parameter. For modified SPRT, this is the targeted alternative toxicity rate p1. For Bayeian Gamma-Poisson model, this is the pair of hyperparameters for the gamma prior on the toxicity rate.}
}
\value{
A list of three: tau, number of events that can trigger a stop, and the corresponding total follow up time.
}
\description{
Internal workhorse function to calculate stopping boundary for a given method for time-to-event data
}
