% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.rule.R
\name{calc.rule}
\alias{calc.rule}
\title{Stopping Rule Calculation}
\usage{
calc.rule(ns, p0, type, param = NULL, alpha, iter = 50)
}
\arguments{
\item{ns}{A vector of sample sizes at which sequential testing is performed}

\item{p0}{The toxicity rate under the null hypothesis}

\item{type}{The method used for constructing the stopping rule}

\item{param}{Extra parameter(s) needed for Bayesian and SPRT methods}

\item{alpha}{The desired type I error / false positive rate for the stopping rule}

\item{iter}{The number of iterations used to search for the boundary}
}
\value{
A matrix with two columns: first column is the vector of sample sizes \code{ns}, second column is a corresponding vector of rejection boundaries at these sizes
}
\description{
Calculate a stopping rule for safety monitoring
}
\examples{
calc.rule(ns=1:30,p0=0.15,type="Pocock",alpha=0.10)
}
\references{
Blah
}
