% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findconst.s.R
\name{findconst.s}
\alias{findconst.s}
\title{Search for Calibration Value (Survival Data)}
\usage{
findconst.s(n, tau, p0, type, alpha, param = NULL)
}
\arguments{
\item{n}{Maximum sample size for safety monitoring}

\item{tau}{Observation period}

\item{p0}{The toxicity rate under the null hypothesis}

\item{type}{The method used for constructing the stopping rule}

\item{alpha}{The desired type I error/false positive rate for the stopping rule}

\item{param}{Extra parameter(s) needed for certain stopping rule methods. For Wang-Tsiatis test, this is the Delta parameter. For modified SPRT, this is the targeted alternative toxicity rate p1. For Bayeian Gamma-Poisson model, this is the pair of hyperparameters for the gamma prior on the toxicity rate.}
}
\value{
The calibration constant used for subsequent stopping boundary calculation
}
\description{
Internal workhorse function to calculate the calibration constant value that attains level alpha for given method
}
\examples{
\dontrun{
findconst.s(n = 30, tau = 100, p0 = 0.1, type = "Pocock", alpha = 0.05)
findconst.s(n = 30, tau = 100, p0 = 0.1, type = "Bayesian", alpha = 0.05, param = c(1.05,1000))
}
}
