% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.rule.R
\name{lines.rule}
\alias{lines.rule}
\title{Add Stopping Rule Curve to Current Plot}
\usage{
\method{lines}{rule}(x, ...)
}
\arguments{
\item{x}{A rule object, being a matrix with two columns: the sample sizes at which sequential testing is performed, and their corresponding rejection boundaries}

\item{...}{Other options to be passed to generic \code{lines} function}
}
\description{
Add a stopping rule graphically as a curve on current plot
}
\examples{
# Binomial Pocock test in 50 patient cohort at 10\% level, expected toxicity rate of 20\%
poc_rule = calc.rule(ns=1:50,p0=0.20,type="Pocock",alpha=0.10)
# Bayesian beta-binomial monitoring in 50 patient cohort at 10\% level, expected toxicity rate of 20\%
bb_rule = calc.rule(ns=1:50,p0=0.20,type="BB",alpha=0.10,param=c(1,7))

# Plot stopping boundaries for stopping rules
plot(poc_rule,col="blue")
lines(bb_rule,col="red")
}
