% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.rule.s.R
\name{lines.rule.s}
\alias{lines.rule.s}
\title{Add Stopping Rule Curve to Current Plot (Survival Data)}
\usage{
\method{lines}{rule.s}(x, ...)
}
\arguments{
\item{x}{A rule object calculated by \code{calc.rule.s} function}

\item{...}{Other options to be passed to generic \code{lines} function}
}
\value{
No return value, function solely modifies current plot
}
\description{
Add Stopping Rule Curve to Current Plot (Survival Data)
}
\examples{
\dontrun{
pocock.rule <- calc.rule.s(n = 30, tau = 100, p0 = 0.1, type = "Pocock", alpha = 0.05)
OBF.rule <- calc.rule.s(n = 30, tau = 100, p0 = 0.1, type = "OBF", alpha = 0.05)
}

}
