% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OC.rule.s.R
\name{OC.rule.s}
\alias{OC.rule.s}
\title{Operating Characteristics Function}
\usage{
OC.rule.s(rule, ps)
}
\arguments{
\item{rule}{A rule object calculated by \code{calc.rule.s} function}

\item{ps}{A vector of toxicity rates at which the operating characteristics will be computed}
}
\value{
A matrix with four columns: the toxicity rate \code{ps}, the corresponding
rejection probabilities, the corresponding expected total follow up time, and
the corresponding expected numbers of events
}
\description{
Compute operating characteristics for a stopping rule at a set of toxicity rates.
Characteristics calculated include the overall rejection probability, the expected
number of patients evaluated, and the expected number of events.
}
\examples{
\dontrun{
pocock.rule <- calc.rule.s(n = 30, tau = 100, p0 = 0.1, type = "Pocock", alpha = 0.05)
OC.rule.s(rule = pocock.rule, ps = seq(0.1, 0.5, 0.1))
}
}
