% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OC.rule.tite.R
\name{OC.rule.tite}
\alias{OC.rule.tite}
\title{Operating Characteristics Function (TITE Method)}
\usage{
OC.rule.tite(rule, ps, ps.compt = NULL, MC, tau, A, family = "power", s = 1)
}
\arguments{
\item{rule}{A \code{rule.tite} object with the safety stopping rule for evaluation}

\item{ps}{Vector of cumulative incidence probabilities for toxicity at time \code{tau}}

\item{ps.compt}{Vector of cumulative incidence probabilities for competing risks at time \code{tau} if competing risks are involved. Set to NULL if no competing risks exist.}

\item{MC}{Number of Monte Carlo replicated datasets to simulate}

\item{tau}{Length of observation period}

\item{A}{Length of accrual period}

\item{family}{Event time distribution, choices including Weibull distribution ('weibull') and power family ('power')}

\item{s}{Shape parameter for Weibull distribution or power parameter for power family}
}
\value{
A matrix with columns containing the toxicity probabilities \code{ps}, competing risk probability (0 for survival outcome), the corresponding rejection probabilities, and the corresponding expected number of events. If \code{tau} and \code{A} are also specified, the expected numbers of enrolled patients and the expected calendar time at the point of stopping/study end are also included.
}
\description{
Compute operating characteristics for a stopping rule at a set of toxicity rates. Characteristics calculated include the overall rejection probability, the expected number of patients evaluated, and the expected number of events.
}
\details{
The Weibull family has cumulative distribution functions of the form
\deqn{1- \exp(- \lambda t^s),}
where \eqn{\lambda} is the rate parameter and \code{s} is the shape parameter.
The power family has cumulative distribution functions of the form
\deqn{1- t^s}
where \code{s} is the power parameter.
If \code{tau} and \code{A} are specified, the expected number of events includes events among patients who are still pending evaluation at the time of early stopping, computed under an assumption of a random uniform accrual distribution. Otherwise, only events that occurred prior to stopping are included, as the number of events occurring in pending patients depends on \code{tau} and \code{A}.
}
\examples{
\dontrun{
# Bayesian beta-extended binomial method in 50 patient cohort at 10\% level,
# expected toxicity probability of 20\%
bb_rule = calc.rule.tite(n=50,p0=0.20,alpha=0.10,type="BB",param=c(2,8))

# Compute operating characteristics at toxicity probabilities of 20\%, 25\%, 30\%, 35\%, and 40\%
OC.rule.tite(rule=bb_rule,ps=seq(0.2,0.4,0.05), MC =1000, tau=30,A=730, family = 'weibull', s = 2)
}
}
