% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opchars.R
\name{opchars}
\alias{opchars}
\title{Operating Characteristics Function}
\usage{
opchars(rule, p)
}
\arguments{
\item{rule}{A matrix with two columns: first column is the vector of sample sizes at which the stopping rule is evaluated; second column is a vector of corresponding vector of rejection boundaries at these sizes}

\item{p}{The event probability}
}
\value{
A list with four objects: A vector of rejection probabilities at the evaluations, the power, the expected number of patients evaluated, and the expected number of events
}
\description{
Workhorse function to calculate operating characteristics for a given stopping rule and event rate
}
