% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.bnd.R
\name{smooth.bnd}
\alias{smooth.bnd}
\title{Smooth Stopping Rule Boundary}
\usage{
smooth.bnd(rule)
}
\arguments{
\item{rule}{A matrix with two columns: the vector of sample sizes at which the stopping rule is evaluated, and a corresponding vector of rejection boundaries at these sizes}
}
\value{
A matrix with two columns: the vector of sample sizes at which the stopping rule is evaluated, and a corresponding vector of smoothed rejection boundaries at these sizes
}
\description{
Function to Smooth Stopping RUle Boundary from Step to Piecewise Linear Function to
}
