% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping.prob.R
\name{stopping.prob}
\alias{stopping.prob}
\title{Calculating the stopping probability given a rejection boundary (Survival Data)}
\usage{
stopping.prob(bnd, p)
}
\arguments{
\item{bnd}{A 'bnd' object calculated by \code{calc.bnd.surv} function}

\item{p}{True toxicity probability}
}
\value{
A list of three: stopping probabilities at each stage, total stopping probability, and non-stopping probabilities of each possible number of events at the last stage.
}
\description{
Internal workhouse function to calculate the stopping probability given a rejection boundary for time-to-event data
}
