% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingrule.R
\name{stoppingrule}
\alias{stoppingrule}
\alias{stoppingrule-package}
\title{Create and Evaluate Stopping Rules for Safety Monitoring}
\description{
Provides functions for creating, displaying, and evaluating stopping rules for safety monitoring in clinical studies.
}
\author{
Michael J. Martens \email{mmartens@mcw.edu}
}
