% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.rule.R
\name{table.rule}
\alias{table.rule}
\title{Tabulate Stopping Rule}
\usage{
table.rule(rule)
}
\arguments{
\item{rule}{A matrix with two columns: the vector of sample sizes at which the stopping rule is evaluated, and a corresponding vector of rejection boundaries at these sizes}
}
\value{
A matrix with two columns: first column lists ranges of evaluable patients; the second column lists the corresponding rejection boundaries for these ranges
}
\description{
Summarize a stopping rule in a condensed tabular format
}
\examples{
poc_rule = calc.rule(ns=1:30,p0=0.15,type="Pocock",alpha=0.10)
table.rule(poc_rule)
}
