% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.rule.s.R
\name{table.rule.s}
\alias{table.rule.s}
\title{Tabulate Stopping Rule (Survival data)}
\usage{
table.rule.s(rule)
}
\arguments{
\item{rule}{A 'rule.s' object calculated by \code{calc.rule.s()} function}
}
\value{
A matrix with two columns: total follow up time and their corresponding rejection boundary.
}
\description{
Summarize a stopping rule in a condensed tabular format
}
\examples{
pocock.rule <- calc.rule.s(n = 30, tau = 100, p0 = 0.1, type = "Pocock", alpha = 0.05)
table.rule.s(pocock.rule)

}
