% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.rule.s.R
\name{calc.rule.s}
\alias{calc.rule.s}
\title{Calculate a stopping rule for safety monitoring (Survival data)}
\usage{
calc.rule.s(n, tau, p0, type, param = NULL, alpha)
}
\arguments{
\item{n}{Maximum sample size for safety monitoring}

\item{tau}{Observation period}

\item{p0}{The toxicity rate under the null hypothesis}

\item{type}{The method used for constructing the stopping rule}

\item{param}{Extra parameter(s) needed for certain stopping rule methods. For Wang-Tsiatis test, this is the Delta parameter. For modified SPRT, this is the targeted alternative toxicity rate p1. For Bayeian Gamma-Poisson model, this is the pair of hyperparameters for the gamma prior on the toxicity rate.}

\item{alpha}{The desired type I error/false positive rate for the stopping rule}
}
\value{
A list of four: 1. A matrix with two columns: total follow up time and their corresponding rejection boundary. 2. value tau to be stored for later use. 3. The calibration constant value used for calculation. 4. Stopping probability at each stage.
}
\description{
Calculate a stopping rule for safety monitoring (Survival data)
}
\examples{
\dontrun{
calc.rule.s(n = 30, tau = 100, p0 = 0.1, type = "Pocock", alpha = 0.05)
}
}
