% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opchars.bin.R
\name{opchars.bin}
\alias{opchars.bin}
\title{Operating Characteristics Function (Binary Data)}
\usage{
opchars.bin(rule, p)
}
\arguments{
\item{rule}{A 'rule.bin' object calculated by \code{calc.rule.bin()} function}

\item{p}{The toxicity probability}
}
\value{
A list with the following elements: p, the corresponding rejection probability, and the corresponding expected numbers of evaluated patients and events at the point of stopping/study end
}
\description{
Internal workhorse function to calculate operating characteristics for a given stopping rule and toxicity probability
}
