% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opchars.R
\name{opchars}
\alias{opchars}
\title{Operating Characteristics Function}
\usage{
opchars(rule, p)
}
\arguments{
\item{rule}{A matrix with two columns: the sample sizes at which sequential testing is performed, and their corresponding rejection boundaries}

\item{p}{The event probability}
}
\value{
A list with four objects: the toxicity rate \code{p}, the corresponding rejection probability, the corresponding expected number of evaluated patients, and the corresponding expected number of events
}
\description{
Workhorse function to calculate operating characteristics for a given stopping rule and event rate
}
