% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rule.R
\name{plot.rule}
\alias{plot.rule}
\title{Plot Stopping Rule}
\usage{
\method{plot}{rule}(rule, add = FALSE, ...)
}
\arguments{
\item{rule}{A rule object, being a matrix with two columns: the vector of sample sizes at which the stopping rule is evaluated, and a corresponding vector of rejection boundaries at these sizes}

\item{add}{Indicator (TRUE/FALSE) of whether stopping rule curve should be added to existing plot}

\item{...}{}
}
\description{
Display a stopping rule graphically as a curve in a plot
}
\examples{
poc_rule = calc.rule(ns=1:30,p0=0.15,type="Pocock",alpha=0.10)
plot(poc_rule,col="blue")
}
