% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.bnd.R
\name{smooth.bnd}
\alias{smooth.bnd}
\title{Smooth Stopping Rule Boundary}
\usage{
smooth.bnd(rule)
}
\arguments{
\item{rule}{A matrix with two columns: the sample sizes at which the stopping rule is evaluated and their corresponding rejection boundaries}
}
\value{
A matrix with two columns: the sample sizes at which the stopping rule is evaluated and their corresponding smoothed rejection boundaries
}
\description{
Function to compute smoothed stopping rule boundary. Takes as input a stopping rule computed by \code{calc.rule()}, where the rejection boundary is a step function, and returns a piecewise linear stopping boundary with identical rejection criteria.
}
