% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping.prob.R
\name{stopping.prob}
\alias{stopping.prob}
\title{Calculating the stopping probability given a rejection boundary (Survival Data)}
\usage{
stopping.prob(bnd, p)
}
\arguments{
\item{bnd}{A 'bnd' object calculated by \code{calc.bnd.s} function}

\item{p}{True toxicity rate}
}
\value{
A list of three: stopping probabilities at each stage, total stopping probability, and non-stopping probabilities of each possible number of events at the last stage.
}
\description{
Calculating the stopping probability given a rejection boundary (Survival Data)
}
\examples{
\dontrun{
pocock.bnd <- calc.bnd.s(n = 30, tau = 100, p0 = 0.1, cval = 4.7, type = "Pocock")
stopping.prob(bnd = pocock.bnd, p = 0.1)
}

}
